+++ src/gui/widgets/graph_view/graph_builder.py
@@ -171,6 +171,9 @@ class GraphBuilder:
         net.set_options(
             """
         {
+            "layout": {
+                "randomSeed": 42
+            },
             "physics": {
                 "enabled": true,
                 "barnesHut": {